feat/avaliations: integrations assingmnet
