+++ src/app/(general)/resultados/[id]/page.tsx
@@ -8,52 +8,46 @@ import { LineData, PieData } from "./dataCharts";
 import { BlueButton } from "@/app/components/buttons/button";
 import { avaliation } from "@/app/interfaces/avaliation";
 import ReportPDF from "./pdf";
-import SpeechBubble from "@/app/components/speechBubble/bubble";
 import BlocoResumo from "@/app/components/blocoResumo/BlocoResumo";
 import { useParams } from "next/navigation";
 import api from "@/utils/api";
 import { UserProps } from "@/app/interfaces/user";
+import { SpeechBubble } from "@/app/components/speechBubble/bubble";
+import Perfil from "@/app/components/perfil/perfil";
+import { useFetchAssignments } from "@/app/hooks/useFetchAssignmnet";
+import { assignment } from "@/app/interfaces/assignment";
+import { formatDate } from "@/utils/formatDate";
 
 const certificateList = ["liderança", "soft skills", "comunicação positiva"];
 
-const avaliations: avaliation[] = [
-  {
-    type: "liderança",
-    dateRealization: "01/01/2021",
-    dateClosing: "01/02/2021",
-  },
-  {
-    type: "soft skills",
-    dateRealization: "01/01/2021",
-    dateClosing: "01/02/2021",
-  },
-  {
-    type: "comunicação positiva",
-    dateRealization: "01/01/2021",
-    dateClosing: "01/02/2021",
-  },
-];
+const userId = "clxtlggn60000cvzgissdxodd";
 
 export default function Resultados() {
-  const { id } = useParams();
-
   const [user, setUser] = useState({} as UserProps);
-  const fetchUser = async () => {
-    const response = await api.get(`/user/${id}`);
-    setUser(response.data);
-  };
+  const [manager, setManager] = useState(false);
+  const [avaliations, setAvaliations] = useState([] as assignment[]);
+
+  const { fecthAssignmentsByUser, getAllAssignments } = useFetchAssignments();
 
   useEffect(() => {
+    const fetchUser = async () => {
+      const response = await api.get(`/user/${userId}`);
+      setUser(response.data);
+      setManager(response.data.manager);
+
+      const assignmentsByUser = await fecthAssignmentsByUser(userId);
+      const assignmentIds = assignmentsByUser.map((au) => au.assignmentId);
+      const fetchedAssignments = await getAllAssignments(assignmentIds);
+      setAvaliations(fetchedAssignments);
+    };
+
     fetchUser();
   }, []);
 
-  const [manager, setManager] = useState(false);
-  useEffect(() => {
-    setManager(user.manager);
-  }, [user]);
-
   return (
     <div className={styles.container}>
+      <Perfil picture={user.image} name={user.name} badge={false} />
+
       {manager ? (
         <>
           <h2 className={styles.title}>
@@ -67,7 +61,7 @@ export default function Resultados() {
           />
         </>
       ) : (
-        <SpeechBubble >
+        <SpeechBubble>
           Sua média no último ciclo foi de: 5.0! Baixe o relatório para mais
           detalhes
         </SpeechBubble>
@@ -120,8 +114,12 @@ export default function Resultados() {
                 className={index % 2 === 0 ? styles.lineWhite : styles.lineGrey}
               >
                 <p className={styles.width}>{avaliation.type}</p>
-                <p className={styles.width}>{avaliation.dateRealization}</p>
-                <p className={styles.width}>{avaliation.dateClosing}</p>
+                <p className={styles.width}>
+                  {formatDate(avaliation.dataAnswered)}
+                </p>
+                <p className={styles.width}>
+                  {formatDate(avaliation.dateConcluded)}
+                </p>
                 <BlueButton
                   width="180px"
                   height="30px"
